pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mistral-ocr-zotero"
version = "0.1.0"
description = "Integrate Mistral OCR API with Zotero-MCP for enhanced PDF-to-markdown conversion"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Mattia" }
]
keywords = ["mistral", "ocr", "zotero", "mcp", "pdf", "markdown"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "mistralai>=1.0.0",
    "pyzotero>=1.5.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "markitdown>=0.1.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
]

[project.scripts]
mistral-ocr-server = "mistral_ocr_zotero.server:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.urls]
Homepage = "https://github.com/mattia/mistral-ocr-zotero"
Documentation = "https://github.com/mattia/mistral-ocr-zotero#readme"
Repository = "https://github.com/mattia/mistral-ocr-zotero"

[tool.hatch.build.targets.wheel]
packages = ["src/mistral_ocr_zotero"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
